Switch to YAML for configuration
Configuration files now use YAML instead of the Java Properties format.
The location of the configuration directory is no longer hard-coded.
Instead it is passed a Java system property: `connector.config-dir`. In
this directory the files `application.yml`, `lms-connector.yml` and
`secrets.yml` are read in that order where properties of the later ones
override those of the previous ones. The default configuration is now
stored in `classpath:config/defaults.yml`.